requirements.txt
# 数据获取
baostock>=0.8.9

# 回测框架
backtrader>=1.9.76.123

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 技术指标（可选，需要先安装系统库）
# talib>=0.4.28  # macOS: brew install ta-lib

# 测试
pytest>=7.4.0
pytest-mock>=3.12.0

# AI 评估
requests>=2.31.0
python-dotenv>=1.0.0
